добавляет селектор для определения списка жанров
удаляет старую логику фильтрации фильмов, все работает через reselect
