+++ src/utils.js
@@ -36,8 +36,7 @@ export const extend = (a, b) => {
   return Object.assign({}, a, b);
 };
 
-export const createRatingText = (rating) => {
-
+const createRatingText = (rating) => {
   let ratingText = null;
   switch (true) {
     case (rating <= 3):
@@ -58,3 +57,22 @@ export const createRatingText = (rating) => {
   }
   return ratingText;
 };
+
+export const adaptFilmsToClient = (films) => films.map((data) => ({
+  id: data.id,
+  backgroundColor: data.background_color,
+  background: data.background_image,
+  poster: data.poster_image,
+  frame: data.preview_image,
+  title: data.name,
+  video: data.preview_video_link,
+  genre: data.genre,
+  year: String(data.released),
+  rating: data.rating,
+  ratingText: createRatingText(data.rating),
+  ratings: String(data.scores_count),
+  description: data.description,
+  director: data.director,
+  starring: data.starring,
+  runtime: data.run_time,
+}));
